moving display functions to separate file
